[Domain/Image] entity->embeddable로 수정(Perfume에 embedded) [Domain/Word] count 추가 -> 가중치를 위해 [Domain/Note] * image 삭제(주석처리) * 연관관계 메서드 생성 -> setWord [Domain/Brand] * perfumes 생성 -> Perfume과 다:1 * image 수정 -> embedded * 연관관계 메서드 생성 -> setWord * 연관관계 메서드 생성 -> setBrand * 연관관계 메서드 생성 -> setPerfumeNote
[Repository/BrandRepository] update 메서드 삭제(주석처리)
[Repository/NoteRepository] update 메서드 삭제(주석처리)
[Repository/ImageRepository] 삭제(주석처리)
[Repository/perfumeSearch] 생성 -> PerfumeRepository findAllByWord 메서드를 위해
[Repository/PerfumeRepository]
* update 메서드 삭제(주석처리)
* findAllByWord 작성중
[Service/ImageService] 삭제(주석처리)
[Service/BrandService]
* update 수정,삭제(주석처리)-> 영속성 컨테스트가 변경
* Transactional 속성변경 -> readonly=true (add는 false)

[Controller/BrandController] 삭제(주석처리)
[Controller/WordController] 삭제(주석처리)

+++ Back-End/src/test/java/com/perfume/surfing/service/BrandServiceTest.java
@@ -32,7 +32,7 @@ public class BrandServiceTest {
         brand.setCreated_at(sqlDate());
 
         //when
-        int saveId = brandService.add(brand);
+        Long saveId = brandService.add(brand);
         System.out.println("savedId: " + saveId);
 
         //then
